requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
# MCP SDK - аль нэгийг сонгох:
# mcp==0.9.0
# эсвэл: pip install git+https://github.com/modelcontextprotocol/python-sdk.git

